+++ lib/animate/animation.js
@@ -90,7 +90,7 @@ class FFAnimation extends FFBase {
       })
       .onUpdate(() => {
         if (!this.proxyAttr) return;
-        // console.log(this.parent.id, 'ani.onUpdate', this.proxyAttr, from);
+        console.log(this.parent.id, 'ani.onUpdate', this.proxyAttr, from);
         const { display } = this.parent;
         this.copyAndResetAttr(this.proxyAttr, from);
         display.attr(this.proxyAttr);
@@ -152,13 +152,13 @@ class FFAnimation extends FFBase {
   complement(from, to) {
     const { display } = this.parent;
     for (let key in from) {
-      if (to[key] === undefined) {
+      if (to[key] === undefined && key !== 'type') {
         to[key] = display.getAttr(key);
       }
     }
 
     for (let key in to) {
-      if (from[key] === undefined) {
+      if (from[key] === undefined && key !== 'type') {
         from[key] = display.getAttr(key);
       }
     }
