+++ lib/node/image.js
@@ -12,8 +12,8 @@
  * @class
  */
 const FFNode = require('./node');
-const TimelineUpdate = require('../timeline/update');
 const { Sprite, Texture } = require('inkpaint');
+const { trimEnd } = require('lodash');
 
 class FFImage extends FFNode {
   constructor(conf = { animations: [] }) {
@@ -97,35 +97,16 @@ class FFImage extends FFNode {
    * @public
    */
   start() {
-    if (this.started) return;
-    this.started = true;
+    if (!super.start()) return false;
+    this.initDraw();
+    return true;
+  }
+
+  initDraw() {
     const path = this.getPath();
     const texture = this.getAsset(path);
     this.draw({ texture });
     this.setDisplaySize();
-    this.animations.start();
-    this.addTimelineCallback();
-  }
-
-  /**
-   * Add callback hook
-   * @private
-   */
-   addTimelineCallback() {
-    this.drawing = this.drawing.bind(this);
-    TimelineUpdate.addFrameCallback(this.drawing);
-  }
-
-  /**
-   * Functions for drawing images
-   * @private
-   */
-  async drawing(timeInMs = 0, nextDeltaInMS = 0) {
-    if (!this.parent.started) return;
-    if (nextDeltaInMS === 0) {
-      // seek的时候，需要强制动一下
-      this.animations.start() && timeInMs;
-    }
   }
 
   /**
