+++ lib/node/node.js
@@ -14,6 +14,7 @@
 const min = require('lodash/min');
 const FFBase = require('../core/base');
 const FFAnimations = require('../animate/animations');
+const TimelineUpdate = require('../timeline/update');
 const { DisplayObject, BLEND_MODES } = require('inkpaint');
 
 class FFNode extends FFBase {
@@ -54,6 +55,27 @@ class FFNode extends FFBase {
    */
   createDisplay() {
     this.display = new DisplayObject();
+  }
+
+  /**
+   * Add callback hook
+   * @private
+   */
+   addTimelineCallback() {
+    this.drawing = this.drawing.bind(this);
+    TimelineUpdate.addFrameCallback(this.drawing);
+  }
+
+  /**
+   * Functions for drawing images
+   * @private
+   */
+  async drawing(timeInMs = 0, nextDeltaInMS = 0) {
+    if (!this.parent.started) return;
+    if (nextDeltaInMS === 0) {
+      // seek的时候，需要强制动一下
+      this.animations.start() && timeInMs;
+    }
   }
 
   /**
@@ -310,7 +332,10 @@ class FFNode extends FFBase {
    * @public
    */
   start() {
+    if (this.started) return false;
     this.animations.start();
+    this.addTimelineCallback();
+    return this.started = true;
   }
 
   /**
