+++ lib/node/video.js
@@ -109,16 +109,8 @@ class FFVideo extends FFImage {
     }
   }
 
-  /**
-   * Start rendering
-   * @private
-   */
-  start() {
-    if (this.started) return;
-    this.started = true; // lock for once
+  initDraw() {
     if (this.backendRender) this.drawCoverImage();
-    this.animations.start();
-    this.addTimelineCallback();
   }
 
   /**
